+++ packages/api-client/src/main/event/BackendEvent.ts
@@ -17,6 +17,12 @@
  *
  */
 
-.text-center {
-  text-align: center;
+import {CONVERSATION_EVENT} from './ConversationEvent';
+import {TEAM_EVENT} from './TeamEvent';
+import {USER_EVENT} from './UserEvent';
+
+interface BackendEvent {
+  type: CONVERSATION_EVENT | TEAM_EVENT | USER_EVENT;
 }
+
+export {BackendEvent};