+++ packages/api-client/CHANGELOG.md
@@ -3,6 +3,78 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.6.36"></a>
+## [0.6.36](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/compare/@wireapp/api-client@0.6.35...@wireapp/api-client@0.6.36) (2018-04-13)
+
+
+
+
+**Note:** Version bump only for package @wireapp/api-client
+
+<a name="0.6.35"></a>
+## [0.6.35](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/compare/@wireapp/api-client@0.6.34...@wireapp/api-client@0.6.35) (2018-04-13)
+
+
+
+
+**Note:** Version bump only for package @wireapp/api-client
+
+<a name="0.6.34"></a>
+## [0.6.34](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/compare/@wireapp/api-client@0.6.33...@wireapp/api-client@0.6.34) (2018-04-12)
+
+
+
+
+**Note:** Version bump only for package @wireapp/api-client
+
+<a name="0.6.33"></a>
+## [0.6.33](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/compare/@wireapp/api-client@0.6.32...@wireapp/api-client@0.6.33) (2018-04-10)
+
+
+
+
+**Note:** Version bump only for package @wireapp/api-client
+
+<a name="0.6.32"></a>
+## [0.6.32](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/compare/@wireapp/api-client@0.6.31...@wireapp/api-client@0.6.32) (2018-04-09)
+
+
+
+
+**Note:** Version bump only for package @wireapp/api-client
+
+<a name="0.6.31"></a>
+## [0.6.31](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/compare/@wireapp/api-client@0.6.30...@wireapp/api-client@0.6.31) (2018-04-06)
+
+
+
+
+**Note:** Version bump only for package @wireapp/api-client
+
+<a name="0.6.30"></a>
+## [0.6.30](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/compare/@wireapp/api-client@0.6.29...@wireapp/api-client@0.6.30) (2018-04-06)
+
+
+
+
+**Note:** Version bump only for package @wireapp/api-client
+
+<a name="0.6.29"></a>
+## [0.6.29](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/compare/@wireapp/api-client@0.6.28...@wireapp/api-client@0.6.29) (2018-04-06)
+
+
+
+
+**Note:** Version bump only for package @wireapp/api-client
+
+<a name="0.6.28"></a>
+## [0.6.28](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/compare/@wireapp/api-client@0.6.27...@wireapp/api-client@0.6.28) (2018-04-06)
+
+
+
+
+**Note:** Version bump only for package @wireapp/api-client
+
 <a name="0.6.27"></a>
 ## [0.6.27](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/compare/@wireapp/api-client@0.6.26...@wireapp/api-client@0.6.27) (2018-04-04)