+++ packages/core/CHANGELOG.md
@@ -3,6 +3,119 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.1.55"></a>
+## [2.1.55](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.54...@wireapp/core@2.1.55) (2018-04-13)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
+<a name="2.1.54"></a>
+## [2.1.54](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.53...@wireapp/core@2.1.54) (2018-04-13)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
+<a name="2.1.53"></a>
+## [2.1.53](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.52...@wireapp/core@2.1.53) (2018-04-12)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
+<a name="2.1.52"></a>
+## [2.1.52](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.51...@wireapp/core@2.1.52) (2018-04-12)
+
+
+### Bug Fixes
+
+* **core:** Synchronize clientId ([#479](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/issues/479)) ([81e3b5f](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/commit/81e3b5f))
+
+
+
+
+<a name="2.1.51"></a>
+## [2.1.51](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.50...@wireapp/core@2.1.51) (2018-04-12)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
+<a name="2.1.50"></a>
+## [2.1.50](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.49...@wireapp/core@2.1.50) (2018-04-12)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
+<a name="2.1.49"></a>
+## [2.1.49](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.48...@wireapp/core@2.1.49) (2018-04-10)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
+<a name="2.1.48"></a>
+## [2.1.48](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.47...@wireapp/core@2.1.48) (2018-04-09)
+
+
+### Bug Fixes
+
+* **cli-client:** Access user context from API client ([#466](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/issues/466)) ([470e875](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/commit/470e875))
+
+
+
+
+<a name="2.1.47"></a>
+## [2.1.47](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.46...@wireapp/core@2.1.47) (2018-04-09)
+
+
+### Bug Fixes
+
+* **core:** Load local client if not found on backend ([#462](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/issues/462)) ([e2306df](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/commit/e2306df))
+
+
+
+
+<a name="2.1.46"></a>
+## [2.1.46](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.45...@wireapp/core@2.1.46) (2018-04-06)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
+<a name="2.1.45"></a>
+## [2.1.45](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.44...@wireapp/core@2.1.45) (2018-04-06)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
+<a name="2.1.44"></a>
+## [2.1.44](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.43...@wireapp/core@2.1.44) (2018-04-06)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
+<a name="2.1.43"></a>
+## [2.1.43](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.42...@wireapp/core@2.1.43) (2018-04-06)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
 <a name="2.1.42"></a>
 ## [2.1.42](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.41...@wireapp/core@2.1.42) (2018-04-04)